fix exit-code stdout-truncation and dedup-comment lie
Per rust reviewer:

1. process::exit doesn't flush stdout. Piped consumers (--format=json
   | jq .) lost the JSON tail (sometimes the entire payload) on
   non-zero exit. Refactor: dispatch returns Result<u8>; main does
   stdout().flush() then constructs ExitCode from the returned code.
   ExitCode runs Drop properly so all buffers flush in order.

2. Dedup comment said 'last write wins' but Vec::dedup_by keeps the
   FIRST element in each run. Updated comment to describe actual
   behavior (first-spec-wins, honoring --against order) and noted the
   tradeoff (cross-spec attribution lost). Behavior unchanged — the
   comment was the bug.

Also: exit_code() now returns u8 (matches ExitCode::from). All other
sites carry that through.

+++ src/output.rs
@@ -162,8 +162,8 @@ fn has_skips(stats: &GatherStats) -> bool {
 
 /// Determine the process exit code from the report. v1 rule: any
 /// `high` or `critical` finding closes the gate (exit 1). Otherwise
-/// exit 0. Tool errors get exit 2 from the caller.
-pub(crate) fn exit_code(report: &AuditReport) -> i32 {
+/// exit 0. Tool errors get exit 2 from main.
+pub(crate) fn exit_code(report: &AuditReport) -> u8 {
     let any_blocking = report
         .findings
         .iter()
